@@ -1,6 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
+<<<<<<< HEAD
 // import { type } from 'os';
+=======
+import { FormserviceService } from 'src/app/service/formservice.service';
+>>>>>>> 1c2baecba9fd275bf44eaf816fbd307abf58cb67
 
 @Component({
   selector: 'app-register',
@@ -27,17 +31,27 @@ export class RegisterComponent implements OnInit {
       Validators.minLength(5)
     ])
   })
-  constructor() { }
-
+  constructor(private formservice: FormserviceService) { }
+  id:any;
+  User:any=[];
+  
   ngOnInit(): void {
 
   }
   get form(): any {
     return this.formregister.controls;
   }
-  onSubmit(): void {
-    if (this.formregister.invalid) {
-      return;
+  onSubmit() {
+    alert('ok')
+    if (!this.formregister.invalid) {
+      this.formservice.postUser(this.formregister.value).subscribe((data)=>{
+        this.User = data;
+      })
+      
+      
+    }
+    else {
+
     }
   }
 
